AuthenticationControllerModifiedToConnectToFrontEndFrontEndLogicImplemented
