+++ root-app-frontend/src/Versions/App1.js
@@ -1,9 +1,9 @@
-import React, { useState } from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from './components/Home';
+// import React, { useState } from 'react';
+import { BrowserRouter as Router, Route} from 'react-router-dom';
+// import Home from './components/Home';
 import Login from './components/Login';
 import Register from './components/Register';
-
+import Header from './components/Header';
 import AddSeed from "./components/Seeds/AddSeed";
 import ViewSeed from "./components/Seeds/ViewSeed";
 import AddPlant from "./components/Plants/AddPlant";
@@ -12,18 +12,16 @@ import AddSoil from "./components/Soils/AddSoil";
 import ViewSoil from "./components/Soils/ViewSoil";
 import AddGarden from "./components/Gardens/AddGarden";
 import ViewGarden from "./components/Gardens/ViewGarden";
-import Login from "./components/Login";
-import Register from "./components/Register";
 import AddJournal from "./components/Journal/AddJournal";
 import ViewJournal from "./components/Journal/ViewJournal";
-import Home from "./components/Home";
+
 
 
 const App = () => {
   return (
     <Router>
       <Header />
-      <Route path="/" exact component={Home} />
+      {/* <Route path="/" exact component={Home} /> */}
       <Route path="/login" component={Login} />
       <Route path="/register" component={Register} />
       <Route path="/seeds/add" component={AddSeed} />
