+++ root-app-frontend/src/Versions/Register.js
@@ -1,0 +1,70 @@
+// import React from 'react';
+// import { Grid, Paper,Avatar,TextField, Button, Typography, Link } from '@mui/material';
+// import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+// import Checkbox from '@mui/material/Checkbox';
+// import FormControlLabel from '@mui/material/FormControlLabel';
+
+
+// const Register = () => {};
+
+// export default Register;
+
+
+
+import React, { useState } from 'react';
+
+const Register = () => {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [verifyPassword, setVerifyPassword] = useState('');
+
+  const handleRegister = async () => {
+    try {
+      const response = await fetch('http://localhost:8080/register', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ username, password, verifyPassword }),
+      });
+
+      if (response.ok) {
+        // Registration successful, handle further actions
+        console.log('Registration successful');
+      } else {
+        // Registration failed, handle error
+        console.error('Registration failed');
+      }
+    } catch (error) {
+      console.error('Error during registration:', error);
+    }
+  };
+
+  return (
+    <div>
+      <h2>Register</h2>
+      <input
+        type="text"
+        placeholder="Username"
+        value={username}
+        onChange={(e) => setUsername(e.target.value)}
+      />
+      <input
+        type="password"
+        placeholder="Password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+      />
+      <input
+        type="password"
+        placeholder="Verify Password"
+        value={verifyPassword}
+        onChange={(e) => setVerifyPassword(e.target.value)}
+      />
+      <button onClick={handleRegister}>Register</button>
+    </div>
+  );
+};
+
+export default Register;
+
